+++ src/__stories__/index.ts
@@ -1,3 +1,8 @@
 // NOTE: The ordering of these imports determines the ordering in Storybook
 
+import './CheckboxInput';
+import './IntegerInput';
+import './JsonInput';
+import './MarkdownInput';
 import './stories.tsx';
+import './StringInput';
